The Step-by-Step Process of Getting a Dental X-ray (OPG) at Crystal Radiology. (02) 8897 868
Crystal Radiology is essential in diagnosing and monitoring various dental and orthodontic conditions. One commonly used dental radiographic procedure is the Orthopantomogram or OPG scan. This imaging technique offers a panoramic view of the entire mouth, including the teeth, jaws, and surrounding structures, which is invaluable for diagnosis and treatment planning.